Removed macOS support
This was done primarily because the runner on GitHub's side had an internal error occur in NSMenu.m on line 777. I'm not exactly sure what the cause of this is due to owning a Mac myself, so I can't test this bug. But my previous fix attempt didn't work. So, for now, it will not support macOS.